Move Before,After to utils
